requirements.txt
streamlit
opencv-python
numpy
pillow
scikit-learn
tensorflow
google-generativeai
